Feature asset model

+++ src/AssetManager.Core/SharedKernel/Entity.cs
@@ -6,6 +6,6 @@ namespace AssetManager.Core.SharedKernel
 {
     public class Entity 
     {
-        public Int64 Id { get; set; }
+        public int Id { get; set; }
     }
 }
